Add fixs TC part 3 (Warden)

+++ src/server/game/Warden/WardenCheckMgr.cpp
@@ -59,7 +59,7 @@ void WardenCheckMgr::LoadWardenChecks()
 
     Field* fields = result->Fetch();
 
-    uint32 maxCheckId = fields[0].GetUInt32();
+    uint16 maxCheckId = fields[0].GetUInt16();
 
     CheckStore.resize(maxCheckId + 1);
 
@@ -119,7 +119,7 @@ void WardenCheckMgr::LoadWardenChecks()
 
         if (checkType == MPQ_CHECK || checkType == MEM_CHECK)
         {
-            WardenCheckResult *wr = new WardenCheckResult();
+            WardenCheckResult* wr = new WardenCheckResult();
             wr->Result.SetHexStr(checkResult.c_str());
             int len = checkResult.size() / 2;
             if (wr->Result.GetNumBytes() < len)
